+++ Casks/dojang.rb
@@ -1,9 +1,9 @@
 cask "dojang" do
   arch arm: "aarch64", intel: "x86_64"
 
-  version "0.2.0"
-  sha256 arm:   "20c6c330aa480d2d54949a3379c82bd0c64865039688f37e406abc5620e294d3",
-         intel: "779ce556294193a68198ab49282d05808256908dc1afcdf8e75b555f26c757b2"
+  version "0.2.1"
+  sha256 arm:   "5ec6d9c35def97cb75440cb9b6de08e884f275c72f1b52eb930c74ac7b894ba0",
+         intel: "c89e475712ab4f66569299355febf5e18e510aab7a63365e8d58017feb1041de"
 
   url "https://github.com/dahlia/dojang/releases/download/#{version}/dojang-#{version}-macos-#{arch}.tar.xz"
   name "Dojang (binary)"